Move `assertion-removal` test to a fixture
