+++ fixtures/assertion-removal/.gitignore
