+++ fixtures/assertion-removal/.prettierignore
