+++ fixtures/assertion-removal/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@fixtures/assertion-removal",
+  "private": true,
+  "exports": {
+    "./sku.config.js": "./sku.config.js"
+  },
+  "dependencies": {
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "sku": "workspace:*"
+  }
+}
