+++ fixtures/assertion-removal/sku.config.js
