+++ fixtures/assertion-removal/src/App.test.tsx
