+++ fixtures/assertion-removal/src/App.tsx
