+++ fixtures/assertion-removal/src/client.tsx
