+++ fixtures/assertion-removal/src/server.tsx
